test: break up NUTs
