+++ .github/workflows/test.yml
@@ -21,6 +21,16 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest]
+        command:
+          - 'yarn test:nuts:bulk:export'
+          - 'yarn test:nuts:bulk:import'
+          - 'yarn test:nuts:data:bulk-upsert-delete'
+          - 'yarn test:nuts:data:create'
+          - 'yarn test:nuts:data:query'
+          - 'yarn test:nuts:data:record'
+          - 'yarn test:nuts:data:search'
+          - 'yarn test:nuts:data:tree'
       fail-fast: false
     with:
       os: ${{ matrix.os }}
+      command: ${{ matrix.command }}
