+++ test/commands/data/query/query.nut.ts
